# Conflicts:
#	cms/views/instructor/add_material.ejs

+++ cms/routes/instructor.js
@@ -190,7 +190,7 @@ router.get('/course-details/:id', function(req, res){
                             }
                             else {
 
-                                connection.query("SELECT * FROM ANOUNCEMENT WHERE Course_id = ?  AND Instructor_id = ? ORDER BY -Posted_on",[req.params.id, rows2[0].Instructor_id] ,function(err4,rows4){
+                                connection.query("SELECT * FROM ANOUNCEMENT WHERE Course_id = ?  AND Instructor_id = ? ORDER BY -Posted_on LIMIT 5",[req.params.id, rows2[0].Instructor_id] ,function(err4,rows4){
                                     if(err4) {
                                         console.log(err4);
                                         res.redirect('/instructor/courses/')
@@ -391,7 +391,7 @@ router.get('/course/announcement/:id1/:id2', function(req, res){
                             }
                             else {
 
-                                connection.query("SELECT * FROM ANOUNCEMENT WHERE Course_id = ? AND Instructor_id = ? ORDER BY -Posted_on",[req.params.id1, rows2[0].Instructor_id] ,function(err4,rows4){
+                                connection.query("SELECT * FROM ANOUNCEMENT WHERE Course_id = ? AND Instructor_id = ? ORDER BY -Posted_on LIMIT 5",[req.params.id1, rows2[0].Instructor_id] ,function(err4,rows4){
                                     if(err4) {
                                         console.log(err4);
                                         res.redirect('/instructor/courses/')
@@ -513,7 +513,7 @@ router.get('/course/:course_id/resources/', function(req, res){
                             }
                             else {
 
-                                connection.query("SELECT * FROM ANOUNCEMENT WHERE Course_id = ?  AND Instructor_id = ? ORDER BY -Posted_on",[req.params.course_id, rows2[0].Instructor_id] ,function(err4,rows4){
+                                connection.query("SELECT * FROM ANOUNCEMENT WHERE Course_id = ?  AND Instructor_id = ? ORDER BY -Posted_on LIMIT 5",[req.params.course_id, rows2[0].Instructor_id] ,function(err4,rows4){
                                     if(err4) {
                                         console.log(err4);
                                         res.redirect('/instructor/courses/')
@@ -773,7 +773,7 @@ router.get('/course/:course_id/tests/', function(req, res){
                             }
                             else {
 
-                                connection.query("SELECT * FROM ANOUNCEMENT WHERE Course_id = ?  AND Instructor_id = ? ORDER BY -Posted_on",[req.params.course_id, rows2[0].Instructor_id] ,function(err4,rows4){
+                                connection.query("SELECT * FROM ANOUNCEMENT WHERE Course_id = ?  AND Instructor_id = ? ORDER BY -Posted_on LIMIT 5",[req.params.course_id, rows2[0].Instructor_id] ,function(err4,rows4){
                                     if(err4) {
                                         console.log(err4);
                                         res.redirect('/instructor/courses/')
@@ -844,7 +844,7 @@ router.get('/course/:course_id/tests/:test_id/edit', function(req, res){
                             }
                             else {
 
-                                connection.query("SELECT * FROM ANOUNCEMENT WHERE Course_id = ?  AND Instructor_id = ? ORDER BY -Posted_on",[req.params.course_id, rows2[0].Instructor_id] ,function(err4,rows4){
+                                connection.query("SELECT * FROM ANOUNCEMENT WHERE Course_id = ?  AND Instructor_id = ? ORDER BY -Posted_on LIMIT 5",[req.params.course_id, rows2[0].Instructor_id] ,function(err4,rows4){
                                     if(err4) {
                                         console.log(err4);
                                         res.redirect('/instructor/courses/')
@@ -928,7 +928,7 @@ router.post('/course/:course_id/tests/:test_id/edit', function (req, res) {
                                 }
                                 else {
 
-                                    connection.query("SELECT * FROM ANOUNCEMENT WHERE Course_id = ?  AND Instructor_id = ? ORDER BY -Posted_on",[req.params.course_id, rows2[0].Instructor_id] ,function(err4,rows4){
+                                    connection.query("SELECT * FROM ANOUNCEMENT WHERE Course_id = ?  AND Instructor_id = ? ORDER BY -Posted_on LIMIT 5",[req.params.course_id, rows2[0].Instructor_id] ,function(err4,rows4){
                                         if(err4) {
                                             console.log(err4);
                                             res.redirect('/instructor/courses/')
@@ -1195,44 +1195,38 @@ router.get('/course/:course_id/tests/:test_id/delete', function (req, res) {
                                     }
                                     else {
 
+                                        if(rows4[0].Question_Paper !== null)
                                         fs.unlink(rows4[0].Question_Paper, (err5) => {
                                             if(err5) {
                                                 console.log(err5);
                                                 res.redirect('/instructor/courses/')
 
                                             }
-                                            else {
+                                        });
 
+                                        if(rows4[0].Answer_Key !== null)
+                                        fs.unlink(rows4[0].Answer_Key, (err5) => {
+                                            if(err5) {
+                                                console.log(err5);
+                                                res.redirect('/instructor/courses/')
 
-                                                fs.unlink(rows4[0].Answer_Key, (err5) => {
-                                                    if(err5) {
-                                                        console.log(err5);
-                                                        res.redirect('/instructor/courses/')
-
-                                                    }
-                                                    else {
-
-
-                                                        connection.query("DELETE FROM TESTS WHERE Test_id = ?",[req.params.test_id] ,function(err5,rows5){
-                                                            if(err5) {
-                                                                console.log(err5);
-                                                                res.redirect('/instructor/courses/')
-
-                                                            }
-                                                            else {
-                                                                console.log('File was deleted');
-                                                                var url = '/instructor/course/' + req.params.course_id + '/tests/';
-                                                                res.redirect(url);
-
-                                                            }
-                                                        });
-
-
-
-                                                    }
-                                                });
                                             }
                                         });
+
+                                        connection.query("DELETE FROM TESTS WHERE Test_id = ?",[req.params.test_id] ,function(err5,rows5){
+                                            if(err5) {
+                                                console.log(err5);
+                                                res.redirect('/instructor/courses/')
+
+                                            }
+                                            else {
+                                                console.log('File was deleted');
+                                                var url = '/instructor/course/' + req.params.course_id + '/tests/';
+                                                res.redirect(url);
+
+                                            }
+                                        });
+
 
                                     }
                                 });
@@ -1279,7 +1273,7 @@ router.get('/course/:course_id/tests/:test_id/results', function(req, res){
                             }
                             else {
 
-                                connection.query("SELECT * FROM ANOUNCEMENT WHERE Course_id = ?  AND Instructor_id = ? ORDER BY -Posted_on",[req.params.course_id, rows2[0].Instructor_id] ,function(err4,rows4){
+                                connection.query("SELECT * FROM ANOUNCEMENT WHERE Course_id = ?  AND Instructor_id = ? ORDER BY -Posted_on LIMIT 5",[req.params.course_id, rows2[0].Instructor_id] ,function(err4,rows4){
                                     if(err4) {
                                         console.log(err4);
                                         res.redirect('/instructor/courses/')
@@ -1304,7 +1298,7 @@ router.get('/course/:course_id/tests/:test_id/results', function(req, res){
                                                     else {
 
 
-                                                        connection.query("SELECT * FROM TEST_RESULTS WHERE Student_id IN (SELECT Student_id FROM ENROLLED WHERE Course_id = ?) ORDER BY Student_id",[req.params.course_id] ,function(err7,rows7){
+                                                        connection.query("SELECT A.First_Name,A.Last_Name, A.Middle_Name, A.Student_id, A.Roll_Number, B.Marks_Obtained, B.Test_id FROM (SELECT * FROM STUDENT WHERE Student_id IN (SELECT Student_id FROM ENROLLED WHERE Course_id = ?) ORDER BY Student_id) A LEFT JOIN (SELECT * FROM TEST_RESULTS WHERE Test_id = ?) B ON A.Student_id = B.Student_id",[req.params.course_id, req.params.test_id] ,function(err7,rows7){
                                                             if(err7) {
                                                                 console.log(err7);
                                                                 res.redirect('/instructor/courses/')
@@ -1369,7 +1363,7 @@ router.post('/course/:course_id/tests/:test_id/results', function(req, res){
                             }
                             else {
 
-                                connection.query("SELECT * FROM ANOUNCEMENT WHERE Course_id = ?  AND Instructor_id = ? ORDER BY -Posted_on",[req.params.course_id, rows2[0].Instructor_id] ,function(err4,rows4){
+                                connection.query("SELECT * FROM ANOUNCEMENT WHERE Course_id = ?  AND Instructor_id = ? ORDER BY -Posted_on LIMIT 5",[req.params.course_id, rows2[0].Instructor_id] ,function(err4,rows4){
                                     if(err4) {
                                         console.log(err4);
                                         res.redirect('/instructor/courses/')
@@ -1491,7 +1485,7 @@ router.get('/course/:course_id/assignments/', function(req, res){
                             }
                             else {
 
-                                connection.query("SELECT * FROM ANOUNCEMENT WHERE Course_id = ?  AND Instructor_id = ? ORDER BY -Posted_on",[req.params.course_id, rows2[0].Instructor_id] ,function(err4,rows4){
+                                connection.query("SELECT * FROM ANOUNCEMENT WHERE Course_id = ?  AND Instructor_id = ? ORDER BY -Posted_on LIMIT 5",[req.params.course_id, rows2[0].Instructor_id] ,function(err4,rows4){
                                     if(err4) {
                                         console.log(err4);
                                         res.redirect('/instructor/courses/')
@@ -1758,7 +1752,7 @@ router.get('/course/:course_id/assignment/:assignment_id/submissions', function(
                             }
                             else {
 
-                                connection.query("SELECT * FROM ANOUNCEMENT WHERE Course_id = ?  AND Instructor_id = ? ORDER BY -Posted_on",[req.params.course_id, rows2[0].Instructor_id] ,function(err4,rows4){
+                                connection.query("SELECT * FROM ANOUNCEMENT WHERE Course_id = ?  AND Instructor_id = ? ORDER BY -Posted_on LIMIT 5",[req.params.course_id, rows2[0].Instructor_id] ,function(err4,rows4){
                                     if(err4) {
                                         console.log(err4);
                                         res.redirect('/instructor/courses/')
@@ -1822,7 +1816,7 @@ router.get('/course/:course_id/students', function(req, res){
                             }
                             else {
 
-                                connection.query("SELECT * FROM ANOUNCEMENT WHERE Course_id = ?  AND Instructor_id = ? ORDER BY -Posted_on",[req.params.course_id, rows2[0].Instructor_id] ,function(err4,rows4){
+                                connection.query("SELECT * FROM ANOUNCEMENT WHERE Course_id = ?  AND Instructor_id = ? ORDER BY -Posted_on LIMIT 5",[req.params.course_id, rows2[0].Instructor_id] ,function(err4,rows4){
                                     if(err4) {
                                         console.log(err4);
                                         res.redirect('/instructor/courses/')
@@ -1889,7 +1883,7 @@ router.get('/course/:course_id/qa', function(req, res){
                             }
                             else {
 
-                                connection.query("SELECT * FROM ANOUNCEMENT WHERE Course_id = ?  AND Instructor_id = ? ORDER BY -Posted_on",[req.params.course_id, rows2[0].Instructor_id] ,function(err4,rows4){
+                                connection.query("SELECT * FROM ANOUNCEMENT WHERE Course_id = ?  AND Instructor_id = ? ORDER BY -Posted_on LIMIT 5",[req.params.course_id, rows2[0].Instructor_id] ,function(err4,rows4){
                                     if(err4) {
                                         console.log(err4);
                                         res.redirect('/instructor/courses/')
@@ -2092,7 +2086,7 @@ router.get('/course/:course_id/assignment/:submission_id/submission', function (
                             }
                             else {
 
-                                connection.query("SELECT * FROM ANOUNCEMENT WHERE Course_id = ? ORDER BY -Posted_on",[req.params.course_id] ,function(err4,rows4){
+                                connection.query("SELECT * FROM ANOUNCEMENT WHERE Course_id = ? ORDER BY -Posted_on LIMIT 5",[req.params.course_id] ,function(err4,rows4){
                                     if(err4) {
                                         console.log(err4);
                                         res.redirect('/instructor/courses/')
@@ -2137,7 +2131,55 @@ router.get('/course/:course_id/assignment/:submission_id/submission', function (
     }
 });
 
+router.get('/course/:course_id/announcements/', function(req, res){
+    if(req.user) {
 
+        connection.query("SELECT * FROM USER WHERE id = ?",req.user ,function(err1,rows1){
+            if(err1) {
+                console.log(err1);
+                res.redirect('/users/login')
+
+            }
+            else {
+                connection.query("SELECT * FROM INSTRUCTOR WHERE User_id = ?  ",req.user ,function(err2,rows2){
+                    if(err2) {
+                        console.log(err2);
+                        res.redirect('/users/login')
+
+                    }
+                    else {
+                        connection.query("SELECT * FROM COURSE WHERE Course_id = ?  ",req.params.course_id ,function(err3,rows3){
+                            if(err3) {
+                                console.log(err3);
+                                res.redirect('/instructor/courses/')
+
+                            }
+                            else {
+
+                                connection.query("SELECT * FROM ANOUNCEMENT WHERE Course_id = ?  AND Instructor_id = ? ORDER BY -Posted_on",[req.params.course_id, rows2[0].Instructor_id] ,function(err4,rows4){
+                                    if(err4) {
+                                        console.log(err4);
+                                        res.redirect('/instructor/courses/')
+
+                                    }
+                                    else {
+
+                                        console.log(rows4);
+                                        res.render('instructor/announcements',{'user':rows1[0], 'instructor': rows2[0], 'course': rows3[0], 'announcements': rows4});
+                                    }
+                                });
+                            }
+                        });
+
+                    }
+                });
+            }
+        });
+    }
+    else{
+        return res.redirect('/users/login')
+    }
+});
 
 
 module.exports = router;
